Move monkeypatch to initializers
